Changes to be committed: 	modified:   app.py 	new file:   app_factory.py 	modified:   command/commands.py 	new file:   extensions.py 	new file:   runner.py
